refactor: 영웅 스탯 형식 float -> double 변화에 대한 UI 컴포넌트 맞춤

+++ Assets/Scripts/Systems/Dummy/HeroTest/HeroEntry.cs
@@ -36,12 +36,12 @@ public class HeroEntry : ScriptableObject
             !PlayerInfo.Instance.TryGetHeroData(HeroId, out HeroSaveData saveData))
         {
             return HeroStats
-                .GetStatTable(HeroId)
+                .GetStatTable((int)HeroId)
                 .GetStat(DefaultLevel);
         }
 
         return HeroStats
-            .GetStatTable(HeroId)
+            .GetStatTable((int)HeroId)
             .GetStat(saveData.Level);
     }
 }
